requirements.txt
# virtualenv virtualenv --no-site-packages
# . virtualenv/bin/activate
# pip install -r requirements.txt
# pip freeze > requirements_freeze.txt
# (vigilar) grep install_required external/*/setup.py


# la base
Django < 1.6

# de base para trabajar
Fabric
South
ipython
django-tinymce

# en produccion
gunicorn
psycopg2

# requerido por algo (django?)
pytz

# para usar imagefields
PIL

# para usar django-filebrowser-no-grappelli tuneado para 1.5
# hay muchas ramas del sw que tienen soporte para 1.6 (mirarlas)
git+https://github.com/jose-lpa/django-filebrowser-no-grappelli.git
pillow

# treemenus
git+https://github.com/dev-fjn/django-treemenus#egg=treemenus

# requerido por django-flatpages-i18n
git+https://github.com/dev-fjn/django-flatpages-i18n#egg=flatpages_i18n
django_modeltranslation

# requerido por zinnia
git+https://github.com/dev-fjn/django-blog-zinnia#egg=zinnia
beautifulsoup4>=4.1
django-tagging
django-mptt

# util para trabajar
python-dateutil

# util para depurar
django-debug-toolbar
django-debug-toolbar-extra
django-debug-toolbar-user-panel-fc
django-dtpanel-htmltidy
django-extensions

